requirements.txt
coloredlogs==15.0.1
contourpy==1.2.0
cycler==0.12.1
flatbuffers==23.5.26
fonttools==4.49.0
humanfriendly==10.0
kiwisolver==1.4.5
matplotlib==3.8.3
mpmath==1.3.0
numpy==1.26.4
onnxruntime==1.17.1
opencv-python==4.9.0.80
packaging==23.2
pillow==10.2.0
protobuf==4.25.3
pyparsing==3.1.1
pyreadline3==3.4.1
python-dateutil==2.9.0.post0
six==1.16.0
sympy==1.12
